Remove old files + move app to components
