+++ src/index.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './components/app/app.jsx';
+import '@ya.praktikum/react-developer-burger-ui-components';
+
+const root = ReactDOM.createRoot(
+  document.getElementById('root')
+);
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
